feat: add NEON SURGE, a twin-stick arena shooter example
A complete survival shooter showcasing the engine end to end:
- Menu / arena / game-over scene flow with fade transitions
- Endless wave director with telegraphed spawn portals and four
  enemy archetypes (chaser, weaver, splitter->shards, comet)
- Dash with afterimages and i-frames, power-ups (rapid, triple,
  shield, heal, score gems) with magnet pickups
- Combo multiplier scoring, floating combat text, persistent
  high score
- Animated FBM nebula via a layer shader with typed uniforms,
  three-tier parallax starfield, batched glow sprites
- Continuous/burst particle emitters, camera follow + shake +
  zoom-kick, scene/game event buses, timers and cooldowns
- Procedural textures and fully synthesized SFX: zero asset files

Also fixes a GCC build break in src/camera.cpp (std::cosf/std::sinf
are not portable; use std::cos/std::sin).

+++ src/camera.cpp
@@ -152,10 +152,10 @@ namespace rlge {
     void Camera3DController::move(const float dx, const float dy, const float dz) { move(Vector3{dx, dy, dz}); }
 
     void Camera3DController::orbitTarget(const float yaw, const float pitch, const float radius) {
-        const float cp = std::cosf(pitch);
-        const float sp = std::sinf(pitch);
-        const float cy = std::cosf(yaw);
-        const float sy = std::sinf(yaw);
+        const float cp = std::cos(pitch);
+        const float sp = std::sin(pitch);
+        const float cy = std::cos(yaw);
+        const float sy = std::sin(yaw);
         const Vector3 offset{
             cy * cp * radius,
             sp * radius,
